improve auto map cleanup

+++ modules/scanning/classes/console/Map.class.php
@@ -8,10 +8,9 @@ namespace scanning\console
 			\Tools::deleteDir("documents/cache");
 			\Tools::deleteDir("documents/statistics");
 
-			\MySQL::getDB()->doQuery("TRUNCATE mapsolarsysteminfocache");
 			\MySQL::getDB()->doQuery("TRUNCATE mapwormholecharacterlocations");
 			\MySQL::getDB()->doQuery("TRUNCATE mapnrofjumps");
-			\MYSQL::getDB()->doQuery("DELETE FROM mapwormholejumplog WHERE jumptime < '".date("Y-m-d", mktime(0,0,0,date("m"),date("d")-5,date("Y")))."'");
+			\MYSQL::getDB()->doQuery("DELETE FROM mapwormholejumplog WHERE jumptime < '".date("Y-m-d", mktime(0,0,0,date("m")-2,date("d"),date("Y")))."'");
 
 			return true;
 		}
@@ -19,21 +18,23 @@ namespace scanning\console
 		function doMaintenance()
 		{
 			\AppRoot::setMaxExecTime(9999);
+            \AppRoot::setMaxMemory("2G");
 			$this->cleanupSignatures();
 			$this->cleanupWormholes();
+            return true;
 		}
 
 		function cleanupSignatures()
 		{
 			// Oude signatures opruimen.
 			$cleanupDate = date("Y-m-d H:i:s", mktime(date("H")-1,date("i"),date("s"),date("m"),date("d")-3,date("Y")));
-			while ($results = \MySQL::getDB()->getRows("SELECT * FROM mapsignatures WHERE updatedate < ? AND deleted = 0 AND sigtype != 'pos'", array($cleanupDate)))
+			if ($results = \MySQL::getDB()->getRows("SELECT * FROM mapsignatures WHERE updatedate < ? AND deleted = 0 AND sigtype != 'pos'", array($cleanupDate)))
 			{
 				foreach ($results as $result)
 				{
 					$signature = new \scanning\model\Signature();
 					$signature->load($result);
-					$signature->delete();
+                    $signature->delete();
 				}
 			}
 
@@ -45,27 +46,35 @@ namespace scanning\console
 
 		function cleanupWormholes()
 		{
-			// Oude wormholes opruimen.
-			$cleanupDate = date("Y-m-d H:i:s", mktime(date("H")-1,date("i"),date("s"),date("m"),date("d")-2,date("Y")));
+            $cleanupDate = date("Y-m-d H:i:s", mktime(date("H")-1,date("i"),date("s"),date("m"),date("d")-2,date("Y")));
 
-			while ($results = \MySQL::getDB()->getRows("SELECT * FROM mapwormholes WHERE updatedate < ? AND permanent = 0", array($cleanupDate)))
+			// Oude wormholes opruimen.
+            if ($results = \MySQL::getDB()->getRows("SELECT * FROM mapwormholes WHERE adddate < ?", array($cleanupDate)))
 			{
 				foreach ($results as $result)
 				{
 					$wormhole = new \scanning\model\Wormhole();
 					$wormhole->load($result);
-					$wormhole->delete();
+                    if (!$wormhole->isPermenant())
+					    $wormhole->delete();
 				}
 			}
 
+
 			// Oude connections opruimen
-			while ($results = \MySQL::getDB()->getRows("SELECT * FROM mapwormholeconnections WHERE updatedate < ?", array($cleanupDate)))
+			if ($results = \MySQL::getDB()->getRows("SELECT * FROM mapwormholeconnections WHERE adddate < ?", array($cleanupDate)))
 			{
 				foreach ($results as $result)
 				{
 					$connection = new \scanning\model\Connection();
 					$connection->load($result);
-					$connection->delete();
+
+                    if ($connection->getFromWormhole()->isPermenant())
+                        continue;
+                    if ($connection->getToWormhole()->isPermenant())
+                        continue;
+
+                    $connection->delete();
 				}
 			}
